Support multiple pots
Each pot has its own sensors and a valve. If a pot decides it's time for
watering, it opens its valve and starts the pump. When finished the pump
is stoped and the valve closed.

Pots are checked sequentially. This makes the code easier and prevents
drawing too much current at once.