Wait a short time for valves and the pump
Valves and the pump don't act instantly. This commit adds a delay after
opening and closing valves and turning off a pump. This should
effectively avoid pumping against a completely closed system.

Before this commit a the pump would be activated right after setting the
valve pin to HIGH. This effectivly turns on the pump and the valve in
the same moment, Until the valve did not physically open, the pump would
work against a closed system. This commit adds 0,5 to all physical
actions.