Pause between readings
It turns out that the wires save some energy (capacitive and/or
inductive). This pause should help to remove that saved energy to
Ground. Otherwise it would affect the sensor readings. I.e. the first
reading of a bulk would be very different as subsequent, as it doesn't
have the saved energy in the wires.