Support AM2321 sensor
The AM2321 humidity and temperature sensor uses one I²C address. I have
to multiplex between other sensors. The original idea, was simply
controlling the power of the sensor with the IO expander. While testing
that, it doesn't work reliably.

First, althout the sensor uses way less
than 4mA, I could't source the voltage from the expander, so I had to
sink it. When controlling the power sink, it turns out that the sensor
worked very unreliable. I.e. I had to wait many seconds before getting a
successful read, and also not every read was successful. Also, there was
some signal on the data bus, so I assume it would sink some power
through the I²C wires.

This commit adds a transistor on the data line, which is controlled by
the IO expander. It works like a bus switch. This produces very reliable
results.