Place pot LED on sensor pin
The sensor pin is most of the times LOW, so it could sink the LED's
current. When reading the sensor the LED will turn off. This will also
help as indicator to see if the pin was successfully restored to LOW
after reading the sensor. Currently the PCF8574 lib doesn't pass any
error codes further. If an LED is off, it's a good indicator for an
invalid state.