Fix bus switch
During prototyping I destroyed a bunch of N7000. It seems they're very
sensitive and need a pull down on the Gate. During prototyping I
connected the gate directly to GND or VDD. I don't know if the act of
unplugging the gate manually did destroy them or connecting them
directly. Adding a pull down resistor fixed that issue. However it might
not be necessary when using the port extender. I need to investigate
further to understand why they broke so easily.