Fix bus switch (firmware only)
In order to support same sensors on multiple pots, the I²C bus needs to
be switched on and off on each pot. That's because the sensors share all
the same address. It turns out to work best with a multiplexer, but the
PCB is not yet updated (also I will make some tests with transistors,
which are smaller and cheaper). This commit doesn't contain any PCB
change, but only a firmware change. A Wire.endTransmission() was
necessary after switching the bus on. This is now abstracted away into
the class BusSwitch.