Generalize probed and switched sensors
It turns out that I can't use a humidity sensor to just measure moisture
in the soil. Humidity in soil is always 100%. However in a recent
experiment I measure humidity near soil. Currently it is additionally
30% more humid than air. If this turns out to be feasible, I will need a
reference sensor for the current air humidity. The reference sensor
would be globally accessible on the I2C bus and can be switched directly
from the Arduino. This commit generalizes the previous code to have a
switched sensor controlled by the Arduino.