Blink LEDs on pots
During the pause, the LEDs of a pot do blink. This helps to signal that
the firmware is still running. Also this commit adds the concept of an
LED to the Pot object. I might in a follow up commit use that to signal
that something was wrong with a specific pot (e.g. a sensor reading
failed).